Add refactor tracking and update User table
Added MRMData.refactorlog for tracking database refactoring operations. Renamed `CreateDate` to `CreatedDate` in the `User` table to better reflect its purpose. The `MRMData.sqlproj` file was updated to include a reference to this new log file, enabling the tracking of changes like renaming. The `MRMData.refactorlog` file now documents the rename operation in the `User` table and the initial rename and subsequent revert in the `Product` table, highlighting the project's evolving requirements and decision-making process.