Add generic SaveData method to SqlDataAccess
Added a new generic method `SaveData<T>` to `MRMDataManager.Library` in the `SqlDataAccess.cs` file. This method allows saving or updating data in the database using any type of parameters. It takes a `storedProcedure`, `parameters` of any type, and a `connectionStringName`, retrieves the connection string, establishes a database connection, and executes the stored procedure.