Add LoggedInUserModel to MRMDesktopUI
A new class, LoggedInUserModel, has been added to the MRMDesktopUI.Library project under the Models directory. This class is designed to represent a user who has logged into the system, containing properties for user identification and basic information (Id, FirstName, LastName, EmailAddress) along with a CreatedDate to track when the user was created. This addition introduces a structured way to handle logged-in user data within the application.