Update DB connection to EFData
Updated the namespace in IdentityModels.cs to change the database connection string from "DefaultConnection" to "EFData". This aligns with the modifications in the Web.config file, where the "DefaultConnection" string has been replaced with "EFData", pointing to a new database file (EFData.mdf). These changes ensure the application connects to the correct database.