Enhance DI setup in Bootstrapper.cs
This commit updates the `Bootstrapper.cs` file within the `MRMDesktopUI` namespace to enhance the dependency injection (DI) setup. Specifically, it adds a crucial step of creating and registering a mapper object right after the AutoMapper configuration is established. This mapper is now available application-wide via the DI container (`_container`). Additionally, the DI container itself is registered as an instance within its scope, catering to advanced use cases where resolving the container might be necessary. Following these enhancements, the commit also ensures that `IProductEndpoint` and `ISaleEndPoint` interfaces, along with their implementations (`ProductEndpoint` and `SaleEndPoint`), are correctly registered to be resolved per request, maintaining the integrity and extensibility of the application's architecture.